requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.3
djangorestframework==3.12.2
drf-yasg==1.20.0
python-decouple==3.3
gunicorn==20.0.4
psycopg2-binary==2.8.6
black==20.8b1